Add validation for delete an update

+++ src/stock/stock.router.ts
@@ -4,7 +4,7 @@
 import express, { Request, Response } from 'express';
 import * as StockService from './stock.service';
 import { check, validationResult } from 'express-validator/check';
-import RULES from './validation';
+import { RULES, CHECK_ID } from './validation';
 /**
  * Router Definition
  */
@@ -37,7 +37,6 @@ stockRouter.get('/:id', async (req: Request, res: Response) => {
 });
 // POST items/
 stockRouter.post('/', RULES, async (req: Request, res: Response) => {
-  console.log(req.body);
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(400).jsonp(errors.array());
@@ -49,16 +48,24 @@ stockRouter.post('/', RULES, async (req: Request, res: Response) => {
     res.status(400).send(e.message);
   }
 });
-stockRouter.put('/', async (req: Request, res: Response) => {
+stockRouter.put('/', CHECK_ID, async (req: Request, res: Response) => {
   try {
-    const stock = await StockService.update(req.body.id, req.body.stock);
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).jsonp(errors.array());
+    }
+    const stock = await StockService.update(req.body.id, req.body);
     res.status(200).send({ stock });
   } catch (e) {
     res.status(400).send(e.message);
   }
 });
-stockRouter.delete('/', async (req: Request, res: Response) => {
+stockRouter.delete('/', CHECK_ID, async (req: Request, res: Response) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).jsonp(errors.array());
+    }
     const stock = await StockService.remove(req.body.id);
     res.status(200).send({ stock });
   } catch (e) {
